+++ pytket/binders/circuit/Circuit/main.cpp
@@ -29,12 +29,12 @@
 #include "tket/Circuit/Circuit.hpp"
 #include "tket/Circuit/Command.hpp"
 #include "tket/Circuit/PauliExpBoxes.hpp"
+#include "tket/Circuit/Simulation/CircuitSimulator.hpp"
 #include "tket/Circuit/ToffoliBox.hpp"
 #include "tket/Gate/OpPtrFunctions.hpp"
 #include "tket/Gate/SymTable.hpp"
 #include "tket/Mapping/Verification.hpp"
 #include "tket/Ops/Op.hpp"
-#include "tket/Simulation/CircuitSimulator.hpp"
 #include "tket/Utils/Json.hpp"
 #include "typecast.hpp"