+++ tket/src/Circuit/Boxes.cpp
@@ -205,6 +205,10 @@ Op_ptr ExpBox::transpose() const {
   return std::make_shared<ExpBox>(A_.transpose(), t_);
 }
 
+std::optional<Eigen::MatrixXcd> ExpBox::get_box_unitary() const {
+  return (i_ * t_ * A_).exp();
+}
+
 void ExpBox::generate_circuit() const {
   circ_ = std::make_shared<Circuit>(two_qubit_canonical((i_ * t_ * A_).exp()));
 }
@@ -363,6 +367,14 @@ Op_ptr QControlBox::dagger() const {
 Op_ptr QControlBox::transpose() const {
   const Op_ptr inner_transpose = op_->transpose();
   return std::make_shared<QControlBox>(inner_transpose, n_controls_);
+}
+
+std::optional<Eigen::MatrixXcd> QControlBox::get_box_unitary() const {
+  const unsigned inner_sz = 1u << n_inner_qubits_;
+  const unsigned sz = inner_sz << n_controls_;
+  Eigen::MatrixXcd u = Eigen::MatrixXcd::Identity(sz, sz);
+  u.bottomRightCorner(inner_sz, inner_sz) = op_->get_unitary();
+  return u;
 }
 
 ProjectorAssertionBox::ProjectorAssertionBox(